+++ src/jcon_client_tcp.c
@@ -450,7 +450,7 @@ size_t jcon_client_tcp_sendData(void *ctx, void *data_ptr, size_t data_size)
 
 //------------------------------------------------------------------------------
 //
-void jcon_client_tcp_log(void *ctx, uint8_t log_type, const char *file, const char *function, uint32_t line, const char *fmt, ...)
+void jcon_client_tcp_log(void *ctx, int log_type, const char *file, const char *function, int line, const char *fmt, ...)
 {
   va_list args;
   char buf[2048];
